store common types in common package
