+++ .github/workflows/frontend.yml
@@ -44,7 +44,8 @@ jobs:
           cd frontend
           yarn install --frozen-lockfile --network-timeout 100000
           yarn setup:common
-
+      - name: setup 'common' package
+        run: cd frontend && yarn run setup:common
       - name: Lint
         run: cd frontend && yarn lint:ci
