+++ alerting/.eslintignore
@@ -1,0 +1,1 @@
+db_data/*
