+++ frontend/src/context/anticipatoryAction/AAStormStateSlice/index.ts
@@ -1,10 +1,10 @@
 import { createAsyncThunk, createSlice, PayloadAction } from '@reduxjs/toolkit';
 import { DateItem } from 'config/types';
+import { StormDataResponseBody } from 'prism-common/dist/types/anticipatory-action-storm/reportResponse';
 import type { CreateAsyncThunkTypes, RootState } from '../../store';
 import { AAStormWindStateReports, AnticipatoryActionState } from './types';
 import { parseAndTransformAA } from './utils';
 import { ParsedStormData } from './parsedStormDataTypes';
-import { StormDataResponseBody } from './rawStormDataTypes';
 
 const initialState: AnticipatoryActionState = {
   data: {},
