+++ frontend/src/context/anticipatoryAction/AAStormStateSlice/utils.ts
@@ -1,3 +1,4 @@
+import { StormDataResponseBody } from 'prism-common/dist/types/anticipatory-action-storm/reportResponse';
 import {
   AACategory,
   AACategoryDataToLandfallMap,
@@ -8,7 +9,6 @@ import {
   MergedFeatures,
   ResultType,
 } from './parsedStormDataTypes';
-import { StormDataResponseBody } from './rawStormDataTypes';
 
 const districtNameMapping: { [key: string]: string } = {
   Maganja_Da_Costa: 'Maganja Da Costa',
